Add previous_job_states table and triggers to fill it
I need to have a list of previous states of each job to write a tool for
displaying builds execution history (currently this information is lost
whenever the job is restarted). Ideally I would save this information in
a JSONB column, but we're on postgresql 9.3, so neither JSONB nor
functions to modify it are available.