Add job states table and triggers to populate it
I need to have a list of previous states of each job to write a tool for
displaying builds execution history (currently this information is lost
whenever the job is restarted). This is done by populating the new
job_states table with triggers that run after a state changes in the
jobs table.

This was originally created when I was working on
https://github.com/travis-pro/team-teal/issues/950. The work was never
deployed, because of the parental leave, but now we're getting back to
it as a part of the offsite coding workshops.