Change set_updated_at proc to use statement_timestamp() instead of now()
`now()` gets a time from the beginning of a current transaction and
`statement_timestamp()` takes the time of the current statement. Thus
it's better to use `statement_timestamp()`, because otherwise the
following scenario might happen:

  1. Open a transaction on thrad A
  2. Open a transaction on thrad B
  3. Update a build record on thread B
  4. Update a build record on thread A (now thread A hangs on the
     `UPDATE` statement until the transaction on thread B finishes)
  5. Commit the transaction on thread B
  6. Commit the transaction on thread A

Because `now()` returns the time of a start of a transaction, the
`updated_at` column would be set to a time of a beginning of the
transaction in thread A, even though a record was last modified in
thread B.

At the moment it doesn't really matter that much, because the triggers
using set_updated_at() are only set on builds and jobs and we rather
don't process a single job or build in parallel, but at some point we
may need to install it on other tables as well and then it would be nice
to have it working correctly.