add new index on repositories (owner_name, name)
this is currently the top query at about 15% of all runtime:

    SELECT "id" FROM "repositories" WHERE (("invalidated_at" IS NULL) AND ("owner_name" = ?) AND ("name" = ?))

it is coming from sync, and it is missing a good index it can use.
thus it needs to do a bitmap index scan on:

* `index_repositories_on_name`
* `index_repositories_on_owner_name`

thus potentially having to scan a lot of repos or owners for
common names. it also needs to reference both of those indexes
and still do a filter for invalidated_at in the end.

with this new index it can index scan directly!