Add request_payload_hashes table
When we migrate data from .org to .com it's hard to ensure that we don't
loose any requests and that we don't have duplicates without checking if
a GitHub requests has been already processed. It seems that the only way
to say if the request has been already processed is to check whether
there was already a request with the same payload.

In order to keep payload hashes I decided to create a new table. This is
because we will be using it only during the time of transition between
.org and .com, so it will be much easier to delete it afterwards rather
than remove a field.