Use NOT VALID clause in constraints
We need to add constraints as NOT VALID because otherwise these would be
blocking operations and we would have to put an app into maintenance
mode in order to execute them. NOT VALID will allow to add constraints
even if existing data isn't valid with a constraint. It will start
enforcing the constraints and at a later time we can validate existing
data with:

    ALTER TABLE builds VALIDATE CONSTRAINT fk_builds_on_repository_id