Make branches unique without cleaning up the table (#160)
We have a problem with the code in gatekeeper and possibly API
introducing duplicated branches. We want to add an index on
`(repository_id, name)`, but we need to clean the table first and in
order to do that we need to fix the bug, so no new duplicates are
introduced during index creation. Otherwise we would have to use a
maintenance window for index creation and stop all the apps from
processing, which would be much harder operationally and add a
considerable delay.

In this commit I introduce a simple way to overcome the problem. A new
column `unique_name` is added to branches. The column is set on UPDATEs
and INSERTs to the branches table with a value of the `name` column. An
index is also added on `(repository_id, unique_name) WHERE unique_name
IS NOT NULL`, which enforces an index on any new or updated branches
when `unique_name` is not NULL.

Note that this will work only for newly created branches as existing
branches don't have the `unique_name` column filled. That's fine,
though, because we create duplicate only when creating new branches in
parallel