Add unique index to crons on branch_id (#169)
* Add unique index to crons on branch_id

There is currently an index on `branch_id`, but it's not unique. Ecto
requires this in order to deal with conflicts on upserts. Also, this is
my understanding of what our current business logic is; there are
currently no duplicate branches, which seems to confirm that this is the
desired behavior.

* Constantize generated class name