Update migrations to allow for json type flexibility
The README states that we need to maintain compatibility with
postgres 9.3, however there are three migrations that use the
`jsonb` type. This type was added in 9.4, which means running
these migrations on 9.3 will result in an error.

This updates three older migrations from 2018 to choose the
correct json type depending on the target postgres version.
It will choose 'json' for 9.3 and 'jsonb' for 9.4 or newer.

The helper method to do this already existed and is used
in other migrations. These specific migrations must have been missed.