Allow postgres 9.4 and lower compatibility for CREATE INDEX migrations
There are several migrations that use the syntax 'CREATE INDEX
CONCURRENTLY IF NOT EXISTS'. Unfortunately the 'IF NOT EXISTS'
portion is only available in postgres 9.5 and higher. In order
to maintain compatibility with older versions these migrations
will omit that portion if necessary.

This is a slight behavior change that will result in a failure
if a conflict is detected.