Add a unique_number column, unique index and triggers
We can't add a unique index on a (repository_id, number) columns pair on
the builds table because we have duplicates in the table that we can't
really get rid of. That's why I came up with a way to ensure uniqueness
on the database level without adding a unique index to the number column
itself, but a unique_number column which is set on INSERT and UPDATE to
the builds table.

I used a similar thing when cleaning up the branches, so it's already
proved to work