Change foreign key constraints to add ON DELETE CASCADE
Initially I didn't want to add ON DELETE CASCADE to foreign key
constraints because I was afraid that it might result in unintended
deletion of data. After thinking about it more I realised that it's not
really an issue. We don't do almost any deletes as a part of automatic
flow of the applications, so changing these constraint shouldn't change
how our apps work and then we only do DELETEs manually when users
request GDPR compliant data cleanup. In this case we actually want all
of the data to be deleted.

In this case it also makes it easier to delete stuff when we actually
needed like in soft_delete_repo_data procedure.