Add function most_recent_non_queued_job_ids_for_user_repositories
This first grabs the repositories a user has permissions for, then
fetches the job ids.

The reason the limit is important is that, however many records we want
from the caller's perspective, we need to ask each individual
repository for, as we can't know which repository or repositories will have the
most
recent jobs. So we have to be pessimistic and assume that each
repository could have *all* the most recent n jobs.

These are then added to a single column table that the caller can query
to get the most recent ids (by ordering/limiting them).

Because we call the single query that leverages our new index in a loop,
each call is quite fast and this shows great promise in significantly increasing
performance.