Speed up jobs queries (#222)
* Add index to jobs on repository_id, id desc and (non-queued) state

* Add function most_recent_non_queued_job_ids_for_user_repositories

This first grabs the repositories a user has permissions for, then
fetches the job ids.

The reason the limit is important is that, however many records we want
from the caller's perspective, we need to ask each individual
repository for, as we can't know which repository or repositories will have the
most
recent jobs. So we have to be pessimistic and assume that each
repository could have *all* the most recent n jobs.

These are then added to a single column table that the caller can query
to get the most recent ids (by ordering/limiting them).

Because we call the single query that leverages our new index in a loop,
each call is quite fast and this shows great promise in significantly increasing
performance.

* Update structure.sql to include migration components

* Remove duplicate migration

* Split out functions

* Rename functions and improve implementation

* Add additional indexes/update naming

* Add final index migration

This adds a sorted index by repository_id, which should speed up queries that do not care about the `state` of jobs.

* Change function to return bigint

This handles the discrepancy we have between the id types for the jobs table on org/com. The results of these migrations have already been applied to com production, so these migrations cannot be safely applied to com until the type has been changed on com from int -> bigint.

* Fix type of id column (int -> bigint)

* Remove reference to non-existent index

* Bump work_mem setting for this specific query